Update st.js - Optional function
## Optional
Optional is a function that can be called on a key as follows:
```js
let data = { items: [1, 2, [3, 4], [5, 6, [7, 8] ] ] }
let template = {
  "{{#optional items}}": [{
    "{{#if items.length > 5}}": "{{items}}"
  }]
}
ST.transform(template, data)
```
This results in following JSON:
```json
{ }
```
This is because, the value attached to the key 'items' is empty. 
The functionality is meant to be same as #? operator just for more complicated situations. The key will not be used if the result is:
 - undefined
 - null
 - false
 - empty object
 - empty array